feat(ui-components): migrate banner component (FRON-93)
Originally the Banner would accept values not specifiyng a content component and was initialized
with an empty banner object for its data variable. Now the default value for data is undefined until
a banner is opened. Opening a banner without specifying a component will be ignored since it could
not be displayed.