fix(transloco): sets the language saved in local storage as default active language
This changes the logic to determine the preloaded active language. If present, it will take the
language provided in local storage under a given key. Each app can customize the key where the
language is stored in local storage via an injection token

BREAKING CHANGE: Every app which uses the SharedTranslocoModule needs to add one more parameter:
localStorageKey. This should be the key of the local storage item or undefined if the language is
not persisted in local storage