fix(workspace): fix Storybook deployment via GH actions
Deploying to GitHub Pages does require the use of GitHub Actions. 
This PR moves the building and deploying of Storybook to GitHub Actions and Upgrades Storybook to v8.

Merging into master is required to confirm the new pipeline is working. 

Dropping the StoryBook aspects from the Jenkins Pipeline as well as upgrading SB to v9 is concern of a different PR